Refactoring
Deleted users, created handlers, all register_message_handler are in function main
