+++ .gitignore
@@ -2,3 +2,7 @@
 .idea
 /venv/
 /env/
+src/__pycache__/handlers.cpython-39.pyc
+src/__pycache__/db.cpython-39.pyc
+src/__pycache__/settings.cpython-39.pyc
+src/__pycache__/utils.cpython-39.pyc
