+++ src/app.py
@@ -1,23 +1,20 @@
+import locale
 import logging
+import handlers
 
-from aiogram.types import Message, BotCommand
+from aiogram import Bot, executor
+from aiogram.contrib.middlewares.logging import LoggingMiddleware
+from aiogram.dispatcher import Dispatcher
+from aiogram.utils.executor import start_webhook
+from aiogram.dispatcher.filters import Text
+from aiogram.types import BotCommand
 from settings import BOT_TOKEN, WEBHOOK_URL, WEBHOOK_PATH, WEBHOOK_HOST, WEBAPP_HOST, WEBAPP_PORT, LOCAL_DEV
 from utils import setup_logging
-from db import DBDriver
-
-from users import register_handlers_user
-
-import locale
 
 try:
     locale.setlocale(locale.LC_TIME, 'ru_RU.UTF-8')
 except:
     locale.setlocale(locale.LC_ALL, 'ru_RU.UTF-8')
-
-from aiogram import Bot, types, executor
-from aiogram.contrib.middlewares.logging import LoggingMiddleware
-from aiogram.dispatcher import Dispatcher, FSMContext
-from aiogram.utils.executor import start_webhook
 
 _logger = logging.getLogger(__name__)
 
@@ -42,43 +39,6 @@ async def on_shutdown():
     await bot.close()
 
 
-@dp.message_handler(regexp="https:\/\/lknpd.nalog.ru/api/v1/receipt/\d+/[\w]+/print")
-async def catch_receipt(message: Message):
-    print(message)
-    driver = DBDriver()
-    driver.add_receipt(message)
-    await message.answer("Чек принят!")
-
-
-@dp.message_handler()
-async def catch_other_message(message: Message):
-    await message.answer("Неизвестный тип сообщений. Воспользуйся командой /help")
-
-
-@dp.message_handler(lambda message: message.text == "Доп. информация", command="additional_info")
-async def additional_info(message: types.Message):
-    await message.reply("Тут будет дополнительная информация")
-
-
-# @dp.message_handler(lambda message: message.text == "Выключить бот", commands="stop")
-# async def stop(message: types.Message):
-#     await on_shutdown()
-#     await message.reply("Бот остановлен")
-
-
-@dp.message_handler(commands="start")
-async def cmd_start(message: types.Message):
-    keyboard = types.InlineKeyboardMarkup()
-    buttons = ["Зарегистрироваться", "Доп. информация", "Выключить бот"]
-    keyboard.add(*buttons)
-    await message.answer("Добро полажловать в TaxBot!", reply_markup=keyboard)
-
-
-@dp.message_handler(commands="help")
-async def get_help_command(message: types.Message):
-    await message.reply("Напиши мне что-нибудь")
-
-
 async def set_commands(bot: Bot):
     commands = [
         BotCommand(command="/help", description="Помощь"),
@@ -93,8 +53,25 @@ def main():
     setup_logging()
     if LOCAL_DEV:
         print("LOCAL development mode!")
-        await set_commands(bot)
-        register_handlers_user(dp)
+        set_commands(bot)
+
+        dp.register_message_handler(handlers.cmd_start, commands="start")
+        dp.register_message_handler(handlers.catch_other_message)
+        dp.register_message_handler(handlers.additional_info, lambda message: message.text == "Доп. информация")
+        dp.register_message_handler(handlers.additional_info, command="additional_info")
+        dp.register_message_handler(handlers.get_help_command, commands="help")
+        dp.register_message_handler(handlers.catch_receipt, regexp="https:\/\/lknpd.nalog.ru/api/v1/receipt/\d+/[\w]+/print")
+
+        dp.register_message_handler(handlers.user_input_start, lambda message: message.text == "Зарегистрироваться")
+        dp.register_message_handler(handlers.user_input_start, commands="registration", state="*")
+        dp.register_message_handler(handlers.user_input_first_name, state=handlers.UserInput.first_name)
+        dp.register_message_handler(handlers.user_input_last_name, state=handlers.UserInput.last_name)
+        dp.register_message_handler(handlers.user_input_patronymic_name, state=handlers.UserInput.patronymic_name)
+        dp.register_message_handler(handlers.user_input_email, state=handlers.UserInput.email)
+        dp.register_message_handler(handlers.cmd_cancel, state=handlers.UserInput.cmd_cancel)
+        dp.register_message_handler(handlers.cmd_cancel, Text(equals="Отмена", ignore_case=True), state="*")
+        dp.register_message_handler(handlers.cmd_cancel, commands="cancel")
+
         executor.start_polling(dp, skip_updates=True)
     else:
         print('Non local run')
